+++ src/components/pokemon/PokemonItem.vue
@@ -35,10 +35,10 @@ import { throttle } from 'lodash';
 import BaseLoader from '@components/ui/BaseLoader.vue';
 import BaseButton from '@components/ui/BaseButton.vue';
 import CenteredColumn from '@components/ui/CenteredColumn.vue';
-import PokemonItemHeader from '@components/pokemons/PokemonItemHeader.vue';
-import PokemonItemStat from '@components/pokemons/PokemonItemStat.vue';
-import PokemonItemType from '@components/pokemons/PokemonItemType.vue';
-import PokemonItemEvolutions from '@components/pokemons/PokemonItemEvolutions.vue';
+import PokemonItemHeader from '@components/pokemon/PokemonItemHeader.vue';
+import PokemonItemStat from '@components/pokemon/PokemonItemStat.vue';
+import PokemonItemType from '@components/pokemon/PokemonItemType.vue';
+import PokemonItemEvolutions from '@components/pokemon/PokemonItemEvolutions.vue';
 import { getPokemonPageBackgroundElement } from '@lib/helpers';
 import { FOURTH_BREAK } from '@constants/resolutions';
 import store from '@lib/store';
