+++ src/components/pokemon/PokemonItemHeader.vue
