+++ src/components/pokemon/PokemonItemType.vue
